requirements.txt
django
psycopg2-binary
djangorestframework
gunicorn
djangorestframework-simplejwt
pillow
drf-yasg
python-decouple

